requirements.txt
Flask
Pillow
flask_cors
ultralytics
matplotlib
numpy
torch
base64
Gunicorn